+++ tests/test_taskhandler_transfer_email.py
@@ -107,6 +107,31 @@ email_plain_dummy_source_task_definition = {
     ],
 }
 
+email_task_definition_message_content_filename = {
+    "type": "transfer",
+    "source": {
+        "hostname": "172.16.0.11",
+        "directory": "/tmp/testFiles/src",
+        "fileRegex": ".*emailhandler.*\\.txt",
+        "protocol": {"name": "ssh", "credentials": {"username": "application"}},
+    },
+    "destination": [
+        {
+            "recipients": ["test@example.com"],
+            "subject": "Test Email Subject",
+            "protocol": {
+                "name": "email",
+                "credentials": {
+                    "username": "{{ lookup('file', path='/tmp/smtp_username') }}",
+                    "password": "{{ lookup('file', path='/tmp/smtp_password') }}",
+                },
+                "sender": "Test Sender <test@example.com>",
+            },
+            "messageContentFilename": "/tmp/email-message.txt",
+        },
+    ],
+}
+
 
 def test_remote_handler():
     # Validate that given a transfer with email protocol, that we get a remote handler of type EmailTransfer
@@ -247,6 +272,130 @@ def test_email_transfer(env_vars, setup_ssh_keys, root_dir):
 
     # Run the transfer
     assert transfer_obj.run()
+
+
+def test_email_transfer_message_content_filename(env_vars, setup_ssh_keys, root_dir):
+
+    import logging
+
+    import opentaskpy.otflogging
+
+    logger = opentaskpy.otflogging.init_logging(
+        __name__, "email-transfer-message-content-filename", level=logging.DEBUG
+    )
+
+    # In GitHub Actions, the variables we need are in the environment
+    # Pull those and write them to the config files first
+    if os.getenv("GITHUB_ACTIONS"):
+        # Get SMTP_USERNAME and SMTP_PASSWORD from environment and write them to files under /tmp
+        fs.create_files(
+            [
+                {
+                    "/tmp/smtp_username": {
+                        "content": os.getenv("SMTP_USERNAME"),
+                    }
+                },
+                {
+                    "/tmp/smtp_password": {
+                        "content": os.getenv("SMTP_PASSWORD"),
+                    }
+                },
+            ]
+        )
+
+    # Create a file to transfer
+    fs.create_files(
+        [{f"{root_dir}/testFiles/ssh_1/src/emailhandler.txt": {"content": "test1234"}}]
+    )
+
+    # Write the email_task_definition to a file which we will read in to resolve the templated values for username and password
+    task_definition_file = (
+        f"{root_dir}/cfg/transfers/email-transfer-message-content-filename.json"
+    )
+    # Delete the file if it exists
+    if os.path.exists(task_definition_file):
+        os.remove(task_definition_file)
+
+    fs.create_files(
+        [
+            {
+                task_definition_file: {
+                    "content": json.dumps(
+                        email_task_definition_message_content_filename
+                    )
+                },
+            },
+            {
+                "/tmp/email-message.txt": {
+                    "content": "This is the email message from a file"
+                },
+            },
+        ]
+    )
+
+    # Load the global config
+    config_loader = ConfigLoader("test/cfg")
+    global_variables = config_loader.get_global_variables()
+    global_variables["global_protocol_vars"] = [
+        {"name": "email", "smtp_port": 587, "smtp_server": "smtp.gmail.com"}
+    ]
+
+    # Load the task definition using the config_loader
+    imported_task_def = config_loader.load_task_definition(
+        "email-transfer-message-content-filename"
+    )
+    os.remove(task_definition_file)
+
+    transfer_obj = transfer.Transfer(
+        global_variables, "email-transfer-message-content-filename", imported_task_def
+    )
+    transfer_obj._set_remote_handlers()
+
+    # Run the transfer
+    assert transfer_obj.run()
+
+    # Check that files have been tidied up on the worker
+    assert not os.path.exists(transfer_obj.local_staging_dir)
+
+    # Check that the message file has been deleted too
+    assert not os.path.exists("/tmp/email-message.txt")
+
+    # Write the email message file again, and this time send the email but explicitl specify not to delete the file
+    fs.create_files(
+        [
+            {
+                "/tmp/email-message.txt": {
+                    "content": "This is the email message from a file"
+                },
+            },
+        ]
+    )
+
+    imported_task_def["destination"][0]["deleteContentFileAfterTransfer"] = False
+    transfer_obj = transfer.Transfer(
+        global_variables, "email-transfer-message-content-filename", imported_task_def
+    )
+    transfer_obj._set_remote_handlers()
+
+    # Run the transfer
+    assert transfer_obj.run()
+
+    # Check that files have been tidied up on the worker
+    assert not os.path.exists(transfer_obj.local_staging_dir)
+
+    # Check that the message file has not been deleted
+    assert os.path.exists("/tmp/email-message.txt")
+
+    # Add a message to the task definition and validate that it failed because the message and messageContentFilename are both defined
+    imported_task_def["destination"][0]["message"] = "This is the email message"
+    transfer_obj = transfer.Transfer(
+        global_variables, "email-transfer-message-content-filename", imported_task_def
+    )
+    transfer_obj._set_remote_handlers()
+
+    # Run the transfer
+    with pytest.raises(Exception) as ex:
+        transfer_obj.run()
 
 
 def test_email_content_type(env_vars, setup_ssh_keys, root_dir):